fix(clippy 1.93): pre-existing + R2.x-introduced lints
Targets the three lints originally flagged in the R2.x P5
discovery (clippy 1.93.0 toolchain update tightened these):

- `kasane-plugin-model/src/lib.rs`: `approx_constant` —
  `SettingValue::Float(3.14)` test literal flagged as a `PI`
  approximation; switch to `2.5` (the test only exercises the
  `Display` impl, the value is incidental).
- `kasane-core/src/widget/backend.rs:364,376`: `unnecessary_cast`
  — `priority: index as i16` / `min_index as i16` where both
  operands are already `i16`. Drop the no-op casts.
- `kasane-core/src/render/halfblock.rs:862`: `useless_vec` —
  `vec![0, 255, 0, 255].repeat(...)` becomes `[0u8, 255, 0, 255]
  .repeat(...)`.

Plus R2.x-introduced cleanup discovered in the same scan:

- Unused `WireFace` / `Style` / `Brush` / `Color` imports across
  `kasane-core/src/{plugin/command, plugin/kakoune_safe_command,
  plugin/render_ornament, plugin/tests/mod, render/inline_decoration,
  render/ornament, state/tests/apply, test_support}.rs` — left
  behind by the migration.
- `kasane-core/src/event_loop/dispatch.rs::Command::RegisterThemeTokens`
  arm: `face.into()` was a `Style → Style` no-op after the type
  change; collapse to direct binding.
- `kasane-core/src/render/grid.rs::default_style` — a TerminalStyle
  helper that became unused in the migration; deleted.
- `kasane-core/benches/{bridge_overhead,iai_pipeline}.rs` and
  `kasane-wasm-bench/benches/{raw_overhead,data_crossing}.rs`:
  stale `face` field names → `style`, `i32 -> i32` casts removed,
  redundant closures collapsed (`b.iter(|| f())` → `b.iter(f)`).

Pre-existing clippy 1.93 noise outside R2.x scope (loop-variable
indexing, items-after-test-module, dead helpers in
`kasane-core/src/state/setting_registry.rs::CoreSettingRegistry::keys`,
etc.) is **not** addressed in this commit — those predate R2.x
and warrant a separate pass focused on broader codebase clippy
hygiene.

+++ kasane-core/src/widget/backend.rs
@@ -361,7 +361,7 @@ mod legacy {
                     let (index, size_hint, element) = matching.into_iter().next().unwrap();
                     Some(Contribution {
                         element,
-                        priority: index as i16,
+                        priority: index,
                         size_hint,
                     })
                 }
@@ -373,7 +373,7 @@ mod legacy {
                         .collect();
                     Some(Contribution {
                         element: Element::row(children),
-                        priority: min_index as i16,
+                        priority: min_index,
                         size_hint: ContribSizeHint::Auto,
                     })
                 }
